+++ Mailer/app/Commands/DemoSeedCommand.php
@@ -8,16 +8,21 @@ use Remp\MailerModule\Repositories\ActiveRow;
 use Remp\MailerModule\Repositories\LayoutsRepository;
 use Remp\MailerModule\Repositories\ListCategoriesRepository;
 use Remp\MailerModule\Repositories\ListsRepository;
+use Remp\MailerModule\Repositories\SnippetsRepository;
 use Remp\MailerModule\Repositories\TemplatesRepository;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 class DemoSeedCommand extends Command
 {
     private $layoutsRepository;
 
     private $templatesRepository;
+
+    private $snippetRepository;
 
     private $listsRepository;
 
@@ -26,12 +31,14 @@ class DemoSeedCommand extends Command
     public function __construct(
         LayoutsRepository $layoutsRepository,
         TemplatesRepository $templatesRepository,
+        SnippetsRepository $snippetRepository,
         ListsRepository $listsRepository,
         ListCategoriesRepository $listCategoriesRepository
     ) {
         parent::__construct();
         $this->layoutsRepository = $layoutsRepository;
         $this->templatesRepository = $templatesRepository;
+        $this->snippetRepository = $snippetRepository;
         $this->listsRepository = $listsRepository;
         $this->listCategoriesRepository = $listCategoriesRepository;
     }
@@ -39,17 +46,59 @@ class DemoSeedCommand extends Command
     protected function configure(): void
     {
         $this->setName('demo:seed')
-            ->setDescription('Seed database with demo values');
+            ->setDescription('Seed database with demo values')
+            ->addArgument(
+                'delete',
+                InputArgument::OPTIONAL,
+                'Remove seed data?',
+                false
+            )
+            ;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
+        $deleteSeedData = $input->getArgument('delete');
+        if ($deleteSeedData === 'delete') {
+            $helper = $this->getHelper('question');
+            $question = new ConfirmationQuestion('Are you sure you want to delete seed data? ', false);
+            if ($helper->ask($input, $output, $question)) {
+                $output->writeln('<info>***** WARNING: Deleting Seed Data *****</info>');
+                $this->deleteSeedData();
+            }
+            return 0;
+        }
+
         $output->writeln('');
         $output->writeln('<info>***** DEMO SEEDER *****</info>');
         $output->writeln('');
 
-        // list category
         $output->write('List categories: ');
+        $category = $this->seedListCategories();
+        $output->writeln('<info>OK!</info>');
+
+        $output->write('Lists: ');
+        $list = $this->seedLists($category);
+        $output->writeln('<info>OK!</info>');
+
+        $output->write('Snippets: ');
+        $snippet = $this->seedSnippets();
+        $output->writeln('<info>OK!</info>');
+
+        $output->write('Layouts: ');
+        $layout = $this->seedLayouts();
+        $output->writeln('<info>OK!</info>');
+
+
+        $output->write('Emails: ');
+        $this->seedEmails($layout, $list);
+        $output->writeln('<info>OK!</info>');
+
+        return Command::SUCCESS;
+    }
+
+    protected function seedListCategories()
+    {
         /** @var ActiveRow $category */
         $category = $this->listCategoriesRepository->findBy('title', 'Newsletters');
         if (!$category) {
@@ -59,10 +108,12 @@ class DemoSeedCommand extends Command
                 'created_at' => new DateTime(),
             ]);
         }
-        $output->writeln('<info>OK!</info>');
 
-        // list
-        $output->write('Lists: ');
+        return $category;
+    }
+
+    protected function seedLists($category)
+    {
         /** @var ActiveRow $list */
         $list = $this->listsRepository->findBy('code', 'demo-weekly-newsletter');
         if (!$list) {
@@ -78,54 +129,102 @@ class DemoSeedCommand extends Command
                 'Example mail list'
             );
         }
-        $output->writeln('<info>OK!</info>');
 
-        // layout
-        $output->write('Layouts: ');
+        return $list;
+    }
+
+
+    protected function seedSnippets()
+    {
+        /** @var ActiveRow $layout */
+        $snippet = $this->snippetRepository->findBy('code', 'demo-snippet');
+        $snippetEmail = $this->snippetRepository->findBy('code', 'demo-snippet-email');
+        if ($snippet && $snippetEmail) {
+            return true;
+        }
+        $snippetText = <<<EOF
+            Remp co.
+            Street 123
+            Lorem City
+            987 654
+            EOF;
+        $snippetHtml = <<<HTML
+            <p>
+                <strong>Remp co.</strong> <br>
+                Street <em>123</em> <br>
+                Lorem City <br>
+                987 654 <br>
+            </p>
+            HTML;
+
+        $emailSnippetText = <<<EOF
+            Written by: John Editor in Chief  Doe
+            EOF;
+        $emailSnippetHtml = <<<HTML
+            <p>Written by: <em>John Editor in Chief Doe</em></p>
+            HTML;
+
+        $snippet = $this->snippetRepository->add('Demo snippet', 'demo-snippet', $snippetText, $snippetHtml);
+        $snippetEmail = $this->snippetRepository->add('Demo Snippet Email', 'demo-snippet-email', $emailSnippetText, $emailSnippetHtml);
+
+        return true;
+    }
+
+    protected function seedLayouts()
+    {
         /** @var ActiveRow $layout */
         $layout = $this->layoutsRepository->findBy('name', 'DEMO layout');
-        if (!$layout) {
-            $text = <<<EOF
-REMP - Reader's Engagement and Monetization Platform
-
-{{ content|raw }}
-
-&copy; REMP, {{ time|date('Y') }}
-EOF;
-            $html = <<<HTML
-<!DOCTYPE html>
-<html lang="en" dir="ltr" xmlns="http://www.w3.org/1999/xhtml">
-    <head>
-    	<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-		<meta charset="utf-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-		<title>REMP - Demo email</title>
-	</head>
-	<body style="-webkit-size-adjust: none; -ms-text-size-adjust: 100%; width: 400px; max-width: 400px;">
-	    <h2>REMP - Reader's Engagement and Monetization Platform</h2>
-        {{ content|raw }}
-        <p><em><small>&copy; REMP, {{ time|date('Y') }}</small></em></p>
-	</body>
-</html>
-HTML;
-            $layout = $this->layoutsRepository->add('DEMO layout', $text, $html);
+        if ($layout) {
+            return $layout;
         }
-        $output->writeln('<info>OK!</info>');
 
-        // email
-        $output->write('Emails: ');
+        $text = <<<EOF
+            REMP - Reader's Engagement and Monetization Platform
+            
+            {{ content|raw }}
+            
+            &copy; REMP, {{ time|date('Y') }}
+            EOF;
+        $html = <<<HTML
+            <!DOCTYPE html>
+            <html lang="en" dir="ltr" xmlns="http://www.w3.org/1999/xhtml">
+                <head>
+                    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+                    <meta charset="utf-8" />
+                    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+                    <title>REMP - Demo email</title>
+                </head>
+                <body style="-webkit-size-adjust: none; -ms-text-size-adjust: 100%; width: 400px; max-width: 400px;">
+                    <h2>REMP - Reader's Engagement and Monetization Platform</h2>
+                    {{ content|raw }}     
+                    <p><em><small>&copy; REMP, {{ time|date('Y') }}</small></em></p>
+                    {{ include('demo-snippet') }}
+                </body>
+            </html>
+            HTML;
+        $layout = $this->layoutsRepository->add('DEMO layout', $text, $html);
+
+        return $layout;
+    }
+
+    protected function seedEmails($layout, $list)
+    {
+        /** @var ActiveRow $email */
         $email = $this->templatesRepository->findBy('code', 'demo-email');
         if (!$email) {
             $text = <<<EOF
-Text content of DEMO email.
-EOF;
+                Text content of DEMO email.
+                EOF;
             $html = <<<HTML
-<table style="width: 400px; background-color: #eee">
-    <tr>
-        <td>Text content of <em>DEMO email</em>.</td>
-    </tr>
-</table>
-HTML;
+                <table style="width: 400px; background-color: #eee">
+                    <tr>
+                        <td>Text content of <em>DEMO email</em>.</td>
+                    </tr>
+                    <tr>
+                        <td>{{ include('demo-snippet-email') }}</td>
+                    </tr>
+                </table>
+                HTML;
 
             $this->templatesRepository->add(
                 'DEMO email',
@@ -139,8 +238,29 @@ HTML;
                 $list->id
             );
         }
-        $output->writeln('<info>OK!</info>');
 
-        return Command::SUCCESS;
+        return $email;
+    }
+
+    protected function deleteSeedData()
+    {
+        if ($email = $this->templatesRepository->findBy('code', 'demo-email')) {
+            $this->templatesRepository->delete($email);
+        }
+        if ($layout = $this->layoutsRepository->findBy('name', 'DEMO layout')) {
+            $this->layoutsRepository->delete($layout);
+        }
+        if ($list = $this->listsRepository->findBy('code', 'demo-weekly-newsletter')) {
+            $this->listsRepository->delete($list);
+        }
+        if ($category = $this->listCategoriesRepository->findBy('title', 'Newsletters')) {
+            $this->listCategoriesRepository->delete($category);
+        }
+        if ($snippet = $this->snippetRepository->findBy('name', 'demo-snippet')) {
+            $this->snippetRepository->delete($snippet);
+        }
+        if ($snippetEmail = $this->snippetRepository->findBy('name', 'demo-snippet-email')) {
+            $this->snippetRepository->delete($snippetEmail);
+        }
     }
 }